+++ src/hardware/vga_pc98_egc.cpp
@@ -99,7 +99,7 @@ void pc98_egc4a0_write(Bitu port,Bitu val,Bitu iolen) {
         return;
     }
 
-    pc98_egc_raw_values[(port>>1u)&7u] = val;
+    pc98_egc_raw_values[(port>>1u)&7u] = (uint16_t)val;
 
     /* assume: (port & 1) == 0 [even] and iolen == 2 */
     switch (port & 0x0E) {
@@ -172,7 +172,7 @@ void pc98_egc4a0_write(Bitu port,Bitu val,Bitu iolen) {
              * If FGC = 1 or BGC = 1:
              *   bits [15:8] = 0
              *   bits [7:0] = foreground color (all 8 bits used in 256-color mode) */
-            pc98_egc_foreground_color = val;
+            pc98_egc_foreground_color = (uint8_t)val;
             pc98_egc_fgcm[0].w = (val & 1) ? 0xFFFF : 0x0000;
             pc98_egc_fgcm[1].w = (val & 2) ? 0xFFFF : 0x0000;
             pc98_egc_fgcm[2].w = (val & 4) ? 0xFFFF : 0x0000;
@@ -180,7 +180,7 @@ void pc98_egc4a0_write(Bitu port,Bitu val,Bitu iolen) {
             break;
         case 0x8: /* 0x4A8 */
             if (pc98_egc_fgc == 0)
-                *((uint16_t*)pc98_egc_mask) = val;
+                *((uint16_t*)pc98_egc_mask) = (uint16_t)val;
             break;
         case 0xA: /* 0x4AA */
             /* If FGC = 0 and BGC = 0:
@@ -188,7 +188,7 @@ void pc98_egc4a0_write(Bitu port,Bitu val,Bitu iolen) {
              * If FGC = 1 or BGC = 1:
              *   bits [15:8] = 0
              *   bits [7:0] = foreground color (all 8 bits used in 256-color mode) */
-            pc98_egc_background_color = val;
+            pc98_egc_background_color = (uint8_t)val;
             pc98_egc_bgcm[0].w = (val & 1) ? 0xFFFF : 0x0000;
             pc98_egc_bgcm[1].w = (val & 2) ? 0xFFFF : 0x0000;
             pc98_egc_bgcm[2].w = (val & 4) ? 0xFFFF : 0x0000;
@@ -253,7 +253,7 @@ void pc98_egc4a0_write_warning(Bitu port,Bitu val,Bitu iolen) {
     switch (port & 0xF) {
         case 0x6:
             /* if the BIOS reports EGC, many early games will write bytewise I/O to port 4A6h */
-            pc98_egc_foreground_color = val;
+            pc98_egc_foreground_color = (uint8_t)val;
             pc98_egc_fgcm[0].w = (val & 1) ? 0xFFFF : 0x0000;
             pc98_egc_fgcm[1].w = (val & 2) ? 0xFFFF : 0x0000;
             pc98_egc_fgcm[2].w = (val & 4) ? 0xFFFF : 0x0000;
